refactor(cas): move the webauthn queries into webauthn/repository.rs
`ceremonies.rs` and `passkeys.rs` now hold only the domain: the ceremony
kinds and states, the credential and its errors. Every query of the
context, the row struct, and the constraint-name mapping behind
`CreateError` live in `repository.rs`, the same shape `accounts` has.
Query text is unchanged, so the offline cache is untouched.
