+++ apps/cas/src/webauthn/ceremonies.rs
@@ -12,16 +12,16 @@
 //! cookie the client could replay. A server-side table is the case its
 //! documentation lists as safe. See `docs/adr/0002-ceremony-state-in-postgres.md`.
 //!
-//! A row is single-use. [`take`] deletes it in the same statement that reads
-//! it, so a challenge answers exactly one request even when two finishes race,
-//! and it is the caller's transaction that decides whether the deletion sticks.
+//! This module is the vocabulary: which ceremonies exist, what state they
+//! carry and what consuming one can yield. Storing and consuming them is
+//! [`repository::start_ceremony`](crate::webauthn::repository::start_ceremony)
+//! and [`repository::take_ceremony`](crate::webauthn::repository::take_ceremony).
 
 use serde::{Serialize, de::DeserializeOwned};
 use uuid::Uuid;
 use webauthn_rs::prelude::PasskeyRegistration;
 
 use crate::accounts::DisplayName;
-use crate::webauthn::{CEREMONY_GRACE, CEREMONY_TIMEOUT};
 
 /// Which ceremony a row belongs to. Finishing looks rows up by kind as well as
 /// by id, so a registration id can never finish a login and vice versa.
@@ -42,9 +42,9 @@ pub trait Ceremony: Serialize + DeserializeOwned {
     const KIND: CeremonyKind;
 }
 
-/// What [`take`] found. A row that was deleted but does not deserialise is not
-/// the same as no row at all: the caller must commit that deletion, or the row
-/// answers the next attempt exactly as badly.
+/// What consuming a ceremony found. A row that was deleted but does not
+/// deserialise is not the same as no row at all: the caller must commit that
+/// deletion, or the row answers the next attempt exactly as badly.
 #[derive(Debug, PartialEq, Eq)]
 pub enum Taken<T> {
     Found(T),
@@ -70,221 +70,4 @@ pub struct PendingRegistration {
 
 impl Ceremony for PendingRegistration {
     const KIND: CeremonyKind = CeremonyKind::Registration;
-}
-
-/// Stores a ceremony and returns its id. The row lives for
-/// [`CEREMONY_TIMEOUT`] plus [`CEREMONY_GRACE`], measured by the database
-/// clock so that every replica agrees on it.
-///
-/// Starting removes nothing: a row past its `expires_at` is ignored by [`take`]
-/// and stays until a separate cleanup deletes it, so housekeeping never sits in
-/// the request path. See `docs/adr/0002-ceremony-state-in-postgres.md`.
-pub async fn start<'e, E, T>(executor: E, state: &T) -> Result<Uuid, sqlx::Error>
-where
-    E: sqlx::PgExecutor<'e>,
-    T: Ceremony,
-{
-    let id = Uuid::new_v4();
-    let state =
-        serde_json::to_value(state).map_err(|error| sqlx::Error::Encode(Box::new(error)))?;
-    let ttl_secs = (CEREMONY_TIMEOUT + CEREMONY_GRACE).as_secs_f64();
-
-    sqlx::query!(
-        r#"INSERT INTO webauthn_ceremonies (id, kind, state, expires_at)
-           VALUES ($1, $2, $3, now() + make_interval(secs => $4))"#,
-        id,
-        T::KIND as CeremonyKind,
-        state,
-        ttl_secs,
-    )
-    .execute(executor)
-    .await?;
-
-    Ok(id)
-}
-
-/// Consumes a ceremony: returns its state and deletes the row in one
-/// statement. Run it inside the transaction that finishes the ceremony, so a
-/// failure later in that transaction leaves the row in place for a retry.
-///
-/// A row whose state no longer deserialises is discarded rather than reported
-/// as an error: it can only be a state shape that changed under a rollout, the
-/// row is worthless either way, and letting the deletion stand is what stops it
-/// from failing every retry until it expires. The caller must commit for that
-/// to hold — see [`Taken::Undecodable`].
-pub async fn take<'e, E, T>(executor: E, id: Uuid) -> Result<Taken<T>, sqlx::Error>
-where
-    E: sqlx::PgExecutor<'e>,
-    T: Ceremony,
-{
-    let row = sqlx::query!(
-        r#"DELETE FROM webauthn_ceremonies
-           WHERE id = $1 AND kind = $2 AND expires_at > now()
-           RETURNING state"#,
-        id,
-        T::KIND as CeremonyKind,
-    )
-    .fetch_optional(executor)
-    .await?;
-
-    let Some(row) = row else {
-        return Ok(Taken::Missing);
-    };
-
-    match serde_json::from_value(row.state) {
-        Ok(state) => Ok(Taken::Found(state)),
-        Err(error) => {
-            tracing::warn!(
-                ceremony_id = %id,
-                error = %error,
-                "discarding a ceremony whose stored state no longer deserialises"
-            );
-            Ok(Taken::Undecodable)
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use sqlx::{PgPool, postgres::types::PgInterval};
-
-    #[derive(Debug, PartialEq, Serialize, serde::Deserialize)]
-    struct State {
-        value: u32,
-    }
-
-    impl Ceremony for State {
-        const KIND: CeremonyKind = CeremonyKind::Registration;
-    }
-
-    /// A state of the other kind, so a mismatch can be provoked without
-    /// passing the kind by hand.
-    #[derive(Debug, PartialEq, Serialize, serde::Deserialize)]
-    struct LoginState {
-        value: u32,
-    }
-
-    impl Ceremony for LoginState {
-        const KIND: CeremonyKind = CeremonyKind::Authentication;
-    }
-
-    /// Moves a ceremony's expiry into the past. Unchecked query: see
-    /// docs/TESTS.md.
-    async fn expire(pool: &PgPool, id: Uuid) {
-        sqlx::query(
-            "UPDATE webauthn_ceremonies SET expires_at = now() - interval '1 second' WHERE id = $1",
-        )
-        .bind(id)
-        .execute(pool)
-        .await
-        .unwrap();
-    }
-
-    async fn count(pool: &PgPool) -> i64 {
-        sqlx::query_scalar("SELECT count(*) FROM webauthn_ceremonies")
-            .fetch_one(pool)
-            .await
-            .unwrap()
-    }
-
-    #[sqlx::test]
-    async fn take_returns_the_state_once(pool: PgPool) {
-        let id = start(&pool, &State { value: 7 }).await.unwrap();
-
-        let first: Taken<State> = take(&pool, id).await.unwrap();
-        let second: Taken<State> = take(&pool, id).await.unwrap();
-
-        assert_eq!(first, Taken::Found(State { value: 7 }));
-        assert_eq!(second, Taken::Missing);
-    }
-
-    #[sqlx::test]
-    async fn take_is_missing_for_an_unknown_id(pool: PgPool) {
-        let state: Taken<State> = take(&pool, Uuid::new_v4()).await.unwrap();
-
-        assert_eq!(state, Taken::Missing);
-    }
-
-    #[sqlx::test]
-    async fn take_is_missing_for_another_kind(pool: PgPool) {
-        let id = start(&pool, &State { value: 7 }).await.unwrap();
-
-        let state: Taken<LoginState> = take(&pool, id).await.unwrap();
-
-        assert_eq!(state, Taken::Missing);
-        // Still there for the right kind: the mismatch consumed nothing.
-        assert_eq!(count(&pool).await, 1);
-    }
-
-    #[sqlx::test]
-    async fn an_expired_ceremony_cannot_be_taken(pool: PgPool) {
-        let id = start(&pool, &State { value: 7 }).await.unwrap();
-        expire(&pool, id).await;
-
-        let state: Taken<State> = take(&pool, id).await.unwrap();
-
-        assert_eq!(state, Taken::Missing);
-    }
-
-    /// A state shape that changed under a rollout: the row is discarded, not
-    /// reported as an error, so it cannot poison every retry until it expires.
-    #[sqlx::test]
-    async fn an_undecodable_state_is_discarded(pool: PgPool) {
-        let id = Uuid::new_v4();
-        // Unchecked query: see docs/TESTS.md.
-        sqlx::query(
-            "INSERT INTO webauthn_ceremonies (id, kind, state, expires_at) VALUES ($1, 'registration', '{\"nope\": 1}'::jsonb, now() + interval '5 minutes')",
-        )
-        .bind(id)
-        .execute(&pool)
-        .await
-        .unwrap();
-
-        let taken: Taken<State> = take(&pool, id).await.unwrap();
-
-        assert_eq!(taken, Taken::Undecodable);
-        assert_eq!(count(&pool).await, 0, "the row is gone, not left to rot");
-    }
-
-    /// The row must outlive the challenge the browser is counting down, or the
-    /// server gives up first and the credential the authenticator just created
-    /// is orphaned.
-    #[sqlx::test]
-    async fn a_ceremony_lives_for_the_timeout_plus_the_grace(pool: PgPool) {
-        let id = start(&pool, &State { value: 7 }).await.unwrap();
-
-        // Unchecked query: see docs/TESTS.md.
-        let lifetime: PgInterval = sqlx::query_scalar(
-            "SELECT expires_at - created_at FROM webauthn_ceremonies WHERE id = $1",
-        )
-        .bind(id)
-        .fetch_one(&pool)
-        .await
-        .unwrap();
-
-        assert_eq!(
-            lifetime,
-            PgInterval {
-                months: 0,
-                days: 0,
-                microseconds: (CEREMONY_TIMEOUT + CEREMONY_GRACE).as_micros() as i64,
-            }
-        );
-    }
-
-    /// Inside a transaction, a rollback puts the ceremony back: a finish that
-    /// fails after consuming the state can be retried.
-    #[sqlx::test]
-    async fn a_rolled_back_take_keeps_the_ceremony(pool: PgPool) {
-        let id = start(&pool, &State { value: 7 }).await.unwrap();
-
-        let mut tx = pool.begin().await.unwrap();
-        let taken: Taken<State> = take(&mut *tx, id).await.unwrap();
-        assert_eq!(taken, Taken::Found(State { value: 7 }));
-        tx.rollback().await.unwrap();
-
-        let again: Taken<State> = take(&pool, id).await.unwrap();
-        assert_eq!(again, Taken::Found(State { value: 7 }));
-    }
 }
